+++ datafusion/execution/src/memory_pool/mod.rs
@@ -23,7 +23,7 @@ use std::{cmp::Ordering, sync::Arc};
 
 mod pool;
 pub mod proxy {
-    pub use datafusion_common::memory_pool::proxy::{RawTableAllocExt, VecAllocExt};
+    pub use datafusion_common::utils::proxy::{RawTableAllocExt, VecAllocExt};
 }
 
 pub use pool::*;
